Refactor: Remove old menu implementation from screen reader
