+++ packages/app-mobile/components/ScreenHeader/index.tsx
@@ -24,7 +24,7 @@ import WarningBanner from './WarningBanner';
 import shim from '@joplin/lib/shim';
 import CommandService from '@joplin/lib/services/CommandService';
 import Icon from '../Icon';
-import DrawerMenu, { MenuOptionType } from './DrawerMenu';
+import Menu, { MenuOptionType } from './Menu';
 export { MenuOptionType };
 
 // Rather than applying a padding to the whole bar, it is applied to each
@@ -697,11 +697,11 @@ class ScreenHeaderComponent extends PureComponent<ScreenHeaderProps, ScreenHeade
 
 		const menuComp =
 			!menuOptions.length || !showContextMenuButton ? null : (
-				<DrawerMenu themeId={this.props.themeId} options={menuOptions}>
+				<Menu themeId={this.props.themeId} options={menuOptions}>
 					<View style={contextMenuStyle}>
 						<Icon name="ionicon ellipsis-vertical" style={this.styles().contextMenuTrigger} accessibilityLabel={_('Actions')}/>
 					</View>
-				</DrawerMenu>
+				</Menu>
 			);
 
 		// Updating the state of this component can result in the left most element becoming hidden, so add a dummy as the first element to prevent this
